adding ai project
